Machine cycle time: allow decimal values (V8 migration + backend + frontend)

+++ tesla-api/src/main/java/com/example/company/machines/adapter/in/web/dto/MachineCreateRequest.java
@@ -11,7 +11,7 @@ public record MachineCreateRequest(
         @Size(max = 20) String code,
         @NotBlank @Size(max = 20) String status,
         @NotBlank @Pattern(regexp = "^(HEADER|LOWER|BOTH)$") String processesType,
-        Integer cycleTimeSeconds,
+        Double cycleTimeSeconds,
         LocalDate lastMaintenanceDate,
         @Size(max = 500) String observations
 ) {
