+++ tesla-api/src/main/java/com/example/company/machines/domain/port/in/UpdateMachineCommand.java
@@ -7,7 +7,7 @@ public record UpdateMachineCommand(
         String code,
         String status,
         String processesType,
-        Integer cycleTimeSeconds,
+        Double cycleTimeSeconds,
         LocalDate lastMaintenanceDate,
         String observations
 ) {
